+++ core/designsystem/src/main/kotlin/com/teamyg/parfait/core/designsystem/component/ygiconbutton/YGIconButtonSize.kt
@@ -1,11 +1,11 @@
-package com.teamyg.parfait.core.designsystem.component.ygbuttonicon
+package com.teamyg.parfait.core.designsystem.component.ygiconbutton
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 
 @Immutable
-enum class YGButtonIconSize(val containerSize: Dp, val iconSize: Dp) {
+enum class YGIconButtonSize(val containerSize: Dp, val iconSize: Dp) {
     SIZE_44(containerSize = 44.dp, iconSize = 24.dp),
     SIZE_48(containerSize = 48.dp, iconSize = 28.dp),
 }
